% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swagger.R
\name{swaggerYaml}
\alias{swaggerYaml}
\title{Generate swagger json from defined functions}
\usage{
swaggerYaml(predict, transform, fit, name, componentVersion)
}
\description{
Generate swagger json from defined functions
}
